backend/services/control-plane/requirements.txt
# Core FastAPI and async
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
uvloop>=0.19.0
httpx>=0.25.0

# Security and auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pynacl>=1.5.0

# Kafka
aiokafka>=0.10.0
kafka-python>=2.0.2

# Redis
redis>=5.0.0
aioredis>=2.0.1

# ClickHouse
clickhouse-connect>=0.6.0
clickhouse-driver>=0.2.6
asynch>=0.2.2

# Protobuf
protobuf>=4.24.0
grpcio-tools>=1.59.0

# Compression
zstandard>=0.22.0
lz4>=4.3.0

# Data processing
numpy>=1.24.0
pandas>=2.1.0
pyarrow>=14.0.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0
treelite>=3.9.0
treelite-runtime>=3.9.0

# Monitoring
prometheus-fastapi-instrumentator>=6.1.0
psutil>=5.9.0

# Database
asyncpg>=0.29.0

# Additional dependencies for complete MEV functionality
websockets>=12.0
python-dateutil>=2.8.2
pydantic[email]>=2.4.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
mypy>=1.7.0
ruff>=0.1.0